[SPARK-54999] Re-enable `ConstraintTests` by moving into `IcebergTests`
